chore: move marketplace sub-components
