+++ autogpt_platform/frontend/src/app/(platform)/admin/marketplace/components/SearchAndFilterForm.tsx
@@ -1,10 +1,7 @@
 "use client";
 
-import { useState, useEffect } from "react";
-import { useRouter, usePathname, useSearchParams } from "next/navigation";
-import { Input } from "@/components/ui/input";
 import { Button } from "@/components/ui/button";
-import { Search } from "lucide-react";
+import { Input } from "@/components/ui/input";
 import {
   Select,
   SelectContent,
@@ -13,13 +10,15 @@ import {
   SelectValue,
 } from "@/components/ui/select";
 import { SubmissionStatus } from "@/lib/autogpt-server-api/types";
+import { Search } from "lucide-react";
+import { usePathname, useRouter, useSearchParams } from "next/navigation";
+import { useEffect, useState } from "react";
 
-export function SearchAndFilterAdminMarketplace({
-  initialSearch,
-}: {
-  initialStatus?: SubmissionStatus;
+interface Props {
   initialSearch?: string;
-}) {
+}
+
+export function SearchAndFilterForm({ initialSearch }: Props) {
   const router = useRouter();
   const pathname = usePathname();
   const searchParams = useSearchParams();
